Feat: Tratando erro 404.

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -17,6 +17,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value = "/topicos")
@@ -41,9 +42,9 @@ public class TopicosController {
     }
 
     @GetMapping(value = "/{id}")
-    public DetalheDoTopicoDto detalhar(@PathVariable Long id){
-        Topico topico = topicoRepository.getOne(id);
-        return new DetalheDoTopicoDto(topico);
+    public ResponseEntity<DetalheDoTopicoDto> detalhar(@PathVariable Long id){
+        Optional<Topico> topico = topicoRepository.findById(id);
+        return topico.isPresent() ? ResponseEntity.ok(new DetalheDoTopicoDto(topico.get())) : ResponseEntity.notFound().build();
     }
 
     @PostMapping
@@ -58,16 +59,25 @@ public class TopicosController {
     @PutMapping("/{id}")
     @Transactional //Avisa o spring que é para ele commitar a transação no final deste metodo
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-        Topico topico = form.atualizar(id,topicoRepository);
-        //Ao final da transação JPA ja detectado que houve alteraçao nas propriedades e ela ja atualiza o banco de dados
+        Optional<Topico> optional = topicoRepository.findById(id);
+        if(optional.isPresent()){
+            Topico topico = form.atualizar(id,topicoRepository);
+            return ResponseEntity.ok(new TopicoDto(topico));
+        }
 
-        return ResponseEntity.ok(new TopicoDto(topico));
+        return ResponseEntity.notFound().build();
+        //Ao final da transação JPA ja detectado que houve alteraçao nas propriedades e ela ja atualiza o banco de dados
     }
 
     @DeleteMapping({"/{id}"})
     @Transactional
     public ResponseEntity<?> delete(@PathVariable Long id){
-        topicoRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        Optional<Topico> optional = topicoRepository.findById(id);
+        if(optional.isPresent()){
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
